pyproject.toml
[project]
name = "ollamaintroduction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub[cli]>=0.34.4",
    "mistral-common>=1.8.4",
    "python-dotenv>=1.1.1",
    "safetensors>=0.6.2",
    "sentencepiece>=0.2.1",
    "torch>=2.8.0",
    "transformers>=4.56.1",
]
